Alex Barcode Scanner - API class implemented

+++ app/src/main/java/com/example/entity/Product.java
@@ -2,7 +2,7 @@ package com.example.entity;
 
 public class Product {
 
-    private int id = 0;
+    private long id = 0;
     private String productName = "default";
     private String productDescription = "default";
     private String barCode = "default";
@@ -11,7 +11,7 @@ public class Product {
     private int quantity = 0;
     private boolean isGlutenFree = false;
 
-    public Product(int id, String productName, String productDescription, String barCode, double price, boolean isGlutenFree){
+    public Product(long id, String productName, String productDescription, String barCode, double price, boolean isGlutenFree){
         setId(id);
         setProductName(productName);
         setProductDescription(productDescription);
@@ -22,12 +22,12 @@ public class Product {
         setDisplayedPrice(price);
     }
 
-    public Product setId(int id){
+    public Product setId(long id){
         this.id = id;
         return this;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
